Refactor reminder reaction message initialisation
